+++ Server/LibSharp/CommonSharp/ResultCode/BRResultCodeGame.cs
@@ -20,73 +20,73 @@ namespace BR {
 	public partial struct ResultCode {
 
 		// User is already in a game 
-		const uint E_GAME_ALREADY_IN_GAME =                                 unchecked(0xE7070000); 
+		const int E_GAME_ALREADY_IN_GAME =                                  unchecked((int)0xE7070000); 
 
 		// Invalid player 
-		const uint E_GAME_INVALID_PLAYER =                                  unchecked(0xE7070001); 
+		const int E_GAME_INVALID_PLAYER =                                   unchecked((int)0xE7070001); 
 
 		// Game is not ready 
-		const uint E_GAME_NOTREADY =                                        unchecked(0xE7070002); 
+		const int E_GAME_NOTREADY =                                         unchecked((int)0xE7070002); 
 
 		// Invalid player role 
-		const uint E_GAME_INVALID_ROLE =                                    unchecked(0xE7070003); 
+		const int E_GAME_INVALID_ROLE =                                     unchecked((int)0xE7070003); 
 
 		// Not ready for next game play state 
-		const uint E_GAME_NOT_READY_FOR_NEXT_STATE =                        unchecked(0xE7070004); 
+		const int E_GAME_NOT_READY_FOR_NEXT_STATE =                         unchecked((int)0xE7070004); 
 
 		// Invalid vote type 
-		const uint E_GAME_INVALID_VOTETYPE =                                unchecked(0xE7070005); 
+		const int E_GAME_INVALID_VOTETYPE =                                 unchecked((int)0xE7070005); 
 
 		// Invalid vote state 
-		const uint E_GAME_INVALID_VOTE_STATE =                              unchecked(0xE7070006); 
+		const int E_GAME_INVALID_VOTE_STATE =                               unchecked((int)0xE7070006); 
 
 		// Invalid vote target 
-		const uint E_GAME_INVALID_VOTE_TARGET =                             unchecked(0xE7070007); 
+		const int E_GAME_INVALID_VOTE_TARGET =                              unchecked((int)0xE7070007); 
 
 		// No suffrage 
-		const uint E_GAME_NO_SUFFRAGE =                                     unchecked(0xE7070008); 
+		const int E_GAME_NO_SUFFRAGE =                                      unchecked((int)0xE7070008); 
 
 		// Invalid game state 
-		const uint E_GAME_INVALID_GAMESTATE =                               unchecked(0xE7070009); 
+		const int E_GAME_INVALID_GAMESTATE =                                unchecked((int)0xE7070009); 
 
 		// Invalid partyID 
-		const uint E_GAME_INVALID_PARTYID =                                 unchecked(0xE707000A); 
+		const int E_GAME_INVALID_PARTYID =                                  unchecked((int)0xE707000A); 
 
 		// Invalid player count for game 
-		const uint E_GAME_INVALID_PLAYER_COUNT =                            unchecked(0xE707000B); 
+		const int E_GAME_INVALID_PLAYER_COUNT =                             unchecked((int)0xE707000B); 
 
 		// User is already in a party 
-		const uint E_GAME_ALREADY_IN_PARTY =                                unchecked(0xE707000C); 
+		const int E_GAME_ALREADY_IN_PARTY =                                 unchecked((int)0xE707000C); 
 
 		// User doesn't have eligible privilege 
-		const uint E_GAME_NO_PRIVILEGE =                                    unchecked(0xE707000D); 
+		const int E_GAME_NO_PRIVILEGE =                                     unchecked((int)0xE707000D); 
 
 		// Invalid shop item data 
-		const uint E_GAME_INVALID_SHOPITEM =                                unchecked(0xE707000E); 
+		const int E_GAME_INVALID_SHOPITEM =                                 unchecked((int)0xE707000E); 
 
 		// Invalid shop item id 
-		const uint E_GAME_INVALID_SHOPITEMID =                              unchecked(0xE707000F); 
+		const int E_GAME_INVALID_SHOPITEMID =                               unchecked((int)0xE707000F); 
 
 		// Not enough money(gold, gem, etc) 
-		const uint E_GAME_NOTENOUGH_RESOURCE =                              unchecked(0xE7070010); 
+		const int E_GAME_NOTENOUGH_RESOURCE =                               unchecked((int)0xE7070010); 
 
 		// Doesn't have enough stamina 
-		const uint E_GAME_LOW_STAMINA =                                     unchecked(0xE7070011); 
+		const int E_GAME_LOW_STAMINA =                                      unchecked((int)0xE7070011); 
 
 		// Invalid player game state 
-		const uint E_GAME_INVALID_PLAYER_STATE =                            unchecked(0xE7070012); 
+		const int E_GAME_INVALID_PLAYER_STATE =                             unchecked((int)0xE7070012); 
 
 		// Maximum notification count 
-		const uint E_GAME_MAX_NOTIFICATION =                                unchecked(0xE7070013); 
+		const int E_GAME_MAX_NOTIFICATION =                                 unchecked((int)0xE7070013); 
 
 		// Maximum try reached 
-		const uint E_GAME_MAX_TRY =                                         unchecked(0xE7070014); 
+		const int E_GAME_MAX_TRY =                                          unchecked((int)0xE7070014); 
 
 		// Invalid cost table id 
-		const uint E_GAME_INVALID_COSTID =                                  unchecked(0xE7070015); 
+		const int E_GAME_INVALID_COSTID =                                   unchecked((int)0xE7070015); 
 
 		// Invalid bot talk table 
-		const uint E_GAME_INVALID_BOTTALK_TABLE =                           unchecked(0xE7070016); 
+		const int E_GAME_INVALID_BOTTALK_TABLE =                            unchecked((int)0xE7070016); 
 
 	}
 }
