+++ Server/LibSharp/CommonSharp/ResultCode/BRResultCodeLogin.cs
@@ -20,19 +20,19 @@ namespace BR {
 	public partial struct ResultCode {
 
 		// Invalid username 
-		const uint E_LOGIN_INVALID_USERNAME =                               unchecked(0xE7060000); 
+		const int E_LOGIN_INVALID_USERNAME =                                unchecked((int)0xE7060000); 
 
 		// Invalid password 
-		const uint E_LOGIN_INVALID_PASSWORD =                               unchecked(0xE7060001); 
+		const int E_LOGIN_INVALID_PASSWORD =                                unchecked((int)0xE7060001); 
 
 		// Already logged in 
-		const uint E_LOGIN_ALREADY_LOGGEDIN =                               unchecked(0xE7060002); 
+		const int E_LOGIN_ALREADY_LOGGEDIN =                                unchecked((int)0xE7060002); 
 
 		// Already logged in at another place, and failed to kick 
-		const uint E_LOGIN_ALREADY_LOGGEDIN_OTHERPLACE =                    unchecked(0xE7060003); 
+		const int E_LOGIN_ALREADY_LOGGEDIN_OTHERPLACE =                     unchecked((int)0xE7060003); 
 
 		// Invalid user signature 
-		const uint E_LOGIN_INVALID_SIGNATURE =                              unchecked(0xE7060004); 
+		const int E_LOGIN_INVALID_SIGNATURE =                               unchecked((int)0xE7060004); 
 
 	}
 }
